refactor(session): extract ActionRouterMixin to _session_actions.py
Behavior-preserving move of dispatch_action, _dispatch_toolbar_action,
_track_action_artifacts, _set_overlay, _set_action_active, _update_vi,
plus the _STAGED_HANDLERS table and _TEST_ACTIONS gating. Constants
re-imported into session.py for back-compat.
